Update Unreal Math includes

+++ Unreal/Plugins/MyPCG/Source/ThirdPartyLibraries/PCGEngine/include/pcg/engine/math/vector3.hpp
@@ -23,9 +23,9 @@ namespace pcg::engine::math
         static const Vector3 forward;
         static const Vector3 backward;
 
-        float x = 0;
-        float y = 0;
-        float z = 0;
+        double x = 0;
+        double y = 0;
+        double z = 0;
 
         /// @brief Get vector containing unit vectors of given axes
         /// @param axes Bit flag representing enabled axes
@@ -40,7 +40,7 @@ namespace pcg::engine::math
 
     Vector3 operator+(const Vector3& lhs, const Vector3& rhs);
     bool operator==(const Vector3& lhs, const Vector3& rhs);
-    Vector3 operator*(const Vector3& vector, float scalar);
+    Vector3 operator*(const Vector3& vector, double scalar);
 }
 
 namespace std
@@ -51,9 +51,9 @@ namespace std
     {
         size_t operator()(const pcg::engine::math::Vector3& vector) const
         {
-            std::size_t x = std::hash<float>{}(vector.x);
-            std::size_t y = std::hash<float>{}(vector.y);
-            std::size_t z = std::hash<float>{}(vector.z);
+            std::size_t x = std::hash<double>{}(vector.x);
+            std::size_t y = std::hash<double>{}(vector.y);
+            std::size_t z = std::hash<double>{}(vector.z);
 
             return x ^ (y << 1) ^ (z << 2);
         }
